+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - Localized export grouping modes
 - Localized inverse option
 - Localized export grouping mode option
+- Export to EDL for importing into DaVinci Resolve (beta, access with `/autocut finish <database> edl`)